libsparse: Use int64_t/off_t for file length/offset
Because the image to be flashed might exceed 2G, even 4G, which will
be out of the range of int or uint type. So we need to use 64-bit int to
indicate the file length. With the new build option "-D_FILE_OFFSET_BITS=64",
off_t will be defined with int64_t.

+++ libsparse/backed_block.c
@@ -25,7 +25,7 @@
 
 struct backed_block {
 	unsigned int block;
-	unsigned int len;
+	int64_t len;
 	enum backed_block_type type;
 	union {
 		struct {
@@ -33,11 +33,11 @@ struct backed_block {
 		} data;
 		struct {
 			char *filename;
-			int64_t offset;
+			off_t offset;
 		} file;
 		struct {
 			int fd;
-			int64_t offset;
+			off_t offset;
 		} fd;
 		struct {
 			uint32_t val;
@@ -62,7 +62,7 @@ struct backed_block *backed_block_iter_next(struct backed_block *bb)
 	return bb->next;
 }
 
-unsigned int backed_block_len(struct backed_block *bb)
+int64_t backed_block_len(struct backed_block *bb)
 {
 	return bb->len;
 }
@@ -90,7 +90,7 @@ int backed_block_fd(struct backed_block *bb)
 	return bb->fd.fd;
 }
 
-int64_t backed_block_file_offset(struct backed_block *bb)
+off_t backed_block_file_offset(struct backed_block *bb)
 {
 	assert(bb->type == BACKED_BLOCK_FILE || bb->type == BACKED_BLOCK_FD);
 	if (bb->type == BACKED_BLOCK_FILE) {
@@ -286,7 +286,7 @@ static int queue_bb(struct backed_block_list *bbl, struct backed_block *new_bb)
 
 /* Queues a fill block of memory to be written to the specified data blocks */
 int backed_block_add_fill(struct backed_block_list *bbl, unsigned int fill_val,
-		unsigned int len, unsigned int block)
+		int64_t len, unsigned int block)
 {
 	struct backed_block *bb = calloc(1, sizeof(struct backed_block));
 	if (bb == NULL) {
@@ -304,7 +304,7 @@ int backed_block_add_fill(struct backed_block_list *bbl, unsigned int fill_val,
 
 /* Queues a block of memory to be written to the specified data blocks */
 int backed_block_add_data(struct backed_block_list *bbl, void *data,
-		unsigned int len, unsigned int block)
+		int64_t len, unsigned int block)
 {
 	struct backed_block *bb = calloc(1, sizeof(struct backed_block));
 	if (bb == NULL) {
@@ -322,7 +322,7 @@ int backed_block_add_data(struct backed_block_list *bbl, void *data,
 
 /* Queues a chunk of a file on disk to be written to the specified data blocks */
 int backed_block_add_file(struct backed_block_list *bbl, const char *filename,
-		int64_t offset, unsigned int len, unsigned int block)
+		off_t offset, int64_t len, unsigned int block)
 {
 	struct backed_block *bb = calloc(1, sizeof(struct backed_block));
 	if (bb == NULL) {
@@ -340,8 +340,8 @@ int backed_block_add_file(struct backed_block_list *bbl, const char *filename,
 }
 
 /* Queues a chunk of a fd to be written to the specified data blocks */
-int backed_block_add_fd(struct backed_block_list *bbl, int fd, int64_t offset,
-		unsigned int len, unsigned int block)
+int backed_block_add_fd(struct backed_block_list *bbl, int fd, off_t offset,
+		int64_t len, unsigned int block)
 {
 	struct backed_block *bb = calloc(1, sizeof(struct backed_block));
 	if (bb == NULL) {
@@ -365,7 +365,7 @@ int backed_block_split(struct backed_block_list *bbl, struct backed_block *bb,
 
 	max_len = ALIGN_DOWN(max_len, bbl->block_size);
 
-	if (bb->len <= max_len) {
+	if ((unsigned int)bb->len <= max_len) {
 		return 0;
 	}
